backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
python-dotenv==1.0.1
slowapi==0.1.9
user-agents==2.2.0
qrcode[pil]==7.4.2

pytest
pytest-asyncio

pytest
pytest-asyncio
httpx
